requirements.txt
Django==4.1.4
djangorestframework==3.14.0
Pillow==9.3.0
django-environ==0.9.0